refactor: make explain.py model, dataset, transformer, and explainer agnostic
