+++ src/xai_lab/utils/paths.py
@@ -47,3 +47,25 @@ def reports_dir_from_run_dir(repo_root: Path, cfg: Dict[str, Any], run_dir: Path
     out = (base_out / "reports" / run_name / run_id).resolve()
     out.mkdir(parents=True, exist_ok=True)
     return out
+
+def resolve_ckpt_and_run_dir(ckpt: Optional[Path], run_dir: Optional[Path]) -> Tuple[Path, Path]:
+    """
+    Decide which checkpoint to load and which run directory it belongs to.
+
+    - If run_dir is provided, use run_dir/best.pt by default.
+    - If ckpt is provided, infer run_dir as ckpt.parent.
+    """
+    if run_dir is not None:
+        run_dir = run_dir.resolve()
+        ckpt_path = (run_dir / "best.pt").resolve()
+        if not ckpt_path.exists():
+            raise FileNotFoundError(f"best.pt not found under run_dir: {ckpt_path}")
+        return ckpt_path, run_dir
+
+    if ckpt is not None:
+        ckpt_path = ckpt.resolve()
+        if not ckpt_path.exists():
+            raise FileNotFoundError(f"Checkpoint not found: {ckpt_path}")
+        return ckpt_path, ckpt_path.parent
+
+    raise ValueError("Provide either --ckpt or --run_dir.")
